sharp framework - added addToTupleSet method to framework - fixed some minor problems with 64bit system compilation - autogen.sh now adds the m4 folder automatically if not existent

+++ src/Problem.cpp
@@ -61,7 +61,7 @@ namespace sharp
 	
 		// when decomposing a bipartite graph
 		if(this->decompositionOptions & BipartiteGraph)
-			ht = be.buildHypertree(hg, BE_MIW_ORDER, (int)this->decompositionParameter);
+			ht = be.buildHypertree(hg, BE_MIW_ORDER, (int)(long)this->decompositionParameter);
 		// when decomposing a default graph
 		else 
 			ht = be.buildHypertree(hg, BE_MIW_ORDER);
@@ -101,7 +101,7 @@ namespace sharp
 	
 	        // when decomposing a bipartite graph
 	        if(this->decompositionOptions & BipartiteGraph)
-	                ht = be.buildHypertree(hg, BE_MIW_ORDER, (int)this->decompositionParameter);
+	                ht = be.buildHypertree(hg, BE_MIW_ORDER, (int)(long)this->decompositionParameter);
 	        // when decomposing a default graph
 	        else
 	                ht = be.buildHypertree(hg, BE_MIW_ORDER);
